moved card specific state properties and methods over to card component
added hooks and am working on being able to update rent duration through lower level datePicker component
